Fix duplicate stitch&cut Textures
It is way more complex than a simple duplicate texture.
There are override of different texture name WITH the SAME texture name. here's a copy paste of my comment:
This is completely fucked up
I hate it so much; Some explaination
0522 Is the locker texture, 64px wide, 32 px tall.
0545 Is the side of that locker texture. 64px wide, 32 px tall.
745 is the full height of the locker texture.
746 is the full height of the side of the locker texture
MIND YOU, ALL OF THIS IS FROM 0545 TEXTURE.
BUT THERE IS AN EXISTING 0522 Texture, WHICH IS A TRIPLE LOCKER , HEIGHT OF 96px.
This whole thing is because we replace a texture with the same name by using the stich/cut method.